+++ sbi/utils/__init__.py
@@ -54,3 +54,4 @@ from sbi.utils.typechecks import (
     is_positive_int,
     is_power_of_two,
 )
+from sbi.utils.sensitivity_analysis import ActiveSubspace