+++ CHANGELOG.md
@@ -7,12 +7,13 @@
 ## API changes
 - `pairplot()`, `conditional_pairplot()`, and `conditional_corrcoeff()` should now be imported from `sbi.analysis` instead of `sbi.utils` (#394).
 - Changed `fig_size` to `figsize` in pairplot (#394).
+- moved `user_input_checks` to `sbi.utils` (#430).
 
 ## Minor changes
 - Depend on new `joblib=1.0.0` and fix progress bar updates for multiprocessing (#421).
 - Fix for embedding nets with `SNRE` (thanks @adittmann, #425)
 - Is it now optional to pass a prior distribution when using SNPE (#426)
-- Support loading of posteriors saved after `sbi v0.13.0` under newer versions (#427)
+- Support loading of posteriors saved after `sbi v0.13.0` under newer versions (#427, thanks @psteinb)
 
 
 # v0.14.3