+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Is it now optional to pass a prior distribution when using SNPE (#426)
 - Support loading of posteriors saved after `sbi v0.15.0` (#427, thanks @psteinb)
 - Neural network training can be resumed (#431)
+- Allow using NSF to estimate 1D distributions (#438)
 
 
 # v0.14.3